requirements.txt
flask==2.3.3
pymongo==4.5.0
python-dotenv==1.0.0
flask-jwt-extended==4.5.2
werkzeug==2.3.7
pytest==7.4.0
requests==2.31.0
python-dateutil==2.8.2
bcrypt==4.0.1